% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detect_random_search}
\alias{detect_random_search}
\title{Detection Function: Law of Random Search}
\usage{
detect_random_search(unit, points, time, g)
}
\arguments{
\item{g}{}
}
\description{
Detection Function: Law of Random Search
}
\examples{

}
\references{
\itemize{
\item \insertRef{Banning2002-uv}{fieldwalkr}
\item \insertRef{Banning2011-wo}{fieldwalkr}
}
}
