% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detect_sweep}
\alias{detect_sweep}
\title{Detection Function: Sweep Width}
\usage{
detect_sweep(unit, points, b, k, r)
}
\arguments{
\item{points}{}
}
\description{
Detection Function: Sweep Width
}
\examples{

}
\references{
\itemize{
\item \insertRef{Banning2011-wo}{fieldwalkr}
}
}
