% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{grab_geoms}
\alias{grab_geoms}
\title{Grab geometries}
\usage{
grab_geoms(mosaic, adjacency, target, set = sample(1:length(adjacency),
  1))
}
\arguments{
\item{adjacency}{A geometric predicate list, see \code{\link[sf:geos_binary_pred]{sf::geos_binary_pred()}}}

\item{target}{Target number of geometries}

\item{set}{Set of geometries already selected}
}
\value{
A list of indexes.
}
\description{
Recursively selects adjacent (etc.) geometries from a geometric predicate
sparse matrix until the target number of polygons is reached.
}
