% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{mosaic}
\alias{mosaic}
\title{Generate a random mosaic}
\usage{
mosaic(frame, density = NULL, area = NULL, method = voronoi)
}
\arguments{
\item{frame}{An \code{sf} object, or an object which can be converted to \code{sf},
containing the sample frame as a spatial polygon}

\item{density}{An integer specifying the desired number of units to be
generated within the sample frame. Not required if \code{area} is
set.}

\item{area}{An integer or double specifying the desired average
area of each unit. The units are assumed to match those
given by \code{\link[sf:st_area]{sf::st_area()}} for \code{frame}. Not required if
\code{density} is set.}

\item{method}{The tessellation model to be used to generate the mosaic;
see details.}
}
\value{
An \code{sf} object containing the units as polygons with an integer \code{id}.
}
\description{
\code{mosaic} generates polygon units within a sample frame using a stochastic
tessellation algorithm. The desired number of units can be specified as an
exact \code{density} or an average unit \code{area}.
}
\details{
\code{method} accepts a function describing the tessalation model to be used (see
\insertCite{Van_Lieshout2012-hb}{fieldwalkr} for an overview). Currently the
only method implemented in the package is \code{\link[=voronoi]{voronoi()}}, which generates a
Voronoi tessellation using a set of uniformly distributed random points and
Delaunay triangulation.
}
\examples{
library(sf)
ashe <- read_sf(system.file("shape/nc.shp", package="sf"))[1,]

# Fixed number of units using Voronoi tesselation
units <- mosaic(ashe, density = 50, method = voronoi)
plot(units)

# Average unit area
library(units)
units <- mosaic(ashe, area = 25000000, method = voronoi)
plot(units)

}
\references{
\insertRef{Van_Lieshout2012-hb}{fieldwalkr}
}
