% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{rpolygon}
\alias{rpolygon}
\title{Random polygon}
\usage{
rpolygon(crs = 3395, origin = c(0, 0), area = 1e+05,
  composite = 64)
}
\arguments{
\item{crs}{An integer EPSG code specifying the coordinate reference system
of the polygon to be generated.
Default: 3395 (World Mercator, https://epsg.io/3395)}

\item{origin}{Numeric. The approximate origin coordinates of the polygon to
be generated (x,y).}

\item{area}{Integer. The approximate area of the polygon to be generated,
in map units (see \code{crs}).}

\item{composite}{Integer. Higher values take longer, but generate more
complex polygons with more sides. See details. Default: 64.}
}
\value{
An \code{sf} object containing a single polygon.
}
\description{
Generates a single, random polygon.
}
\details{
\code{rpolygon()} generates a random tessellation using \code{\link[=mosaic]{mosaic()}} and returns a
random polygon sampled for it. If \code{composite} > 1, several contiguious tiles
are dissolved, creating more complex polygons with more sides.
}
\examples{

# Simple polygon
polygon <- rpolygon()
plot(polygon)

# More complex polygon
polygon <- rpolygon(composite = 8)

}
