% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_random}
\alias{sample_random}
\title{Randomly sample survey units}
\usage{
sample_random(units, n = NULL, fraction = NULL, column = "sample")
}
\arguments{
\item{units}{An \code{sf} object of survey units.}

\item{n}{Number of units to sample. Either \code{n} or \code{fraction} must be
set.}

\item{fraction}{Fraction of units to sample.}

\item{column}{Name of the column to be added. Default: "sample".}
}
\value{
\code{units} with an additional boolean column describing the sample.
}
\description{
\code{sample_random()} generates a simple random sample of a set of survey units.
}
\examples{
library(magrittr)
library(ggplot2)

rpolygon() \%>\%
  quadrats(n = c(10, 10)) \%>\%
  sample_random(n = 50) \%>\%
  ggplot(aes(fill = sample)) + geom_sf()
}
