% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_systematic}
\alias{sample_systematic}
\title{Systematically sample survey units}
\usage{
sample_systematic(units, n, start = 1, column = "sample")
}
\arguments{
\item{units}{An sf object of survey units.}

\item{n}{Step size, i.e. selects every \code{n}th unit.}

\item{start}{Index of the unit to start sampling from.}

\item{column}{Name of the column to be added. Default: "sample".}
}
\value{
\code{units} with an additional boolean column describing the sample.
}
\description{
\code{sample_systematic()} generates a systematic sample of survey units.
}
\examples{
library(magrittr)
library(ggplot2)

rpolygon() \%>\%
  quadrats(n = c(10, 10)) \%>\%
  sample_systematic(n = 10) \%>\%
  ggplot(aes(fill = sample)) + geom_sf()
}
