% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{voronoi}
\alias{voronoi}
\title{Random tessellation using Voronoi tiles}
\usage{
voronoi(poly, n, warn_multipart = TRUE)
}
\arguments{
\item{poly}{An \code{sf} object, or an object that can be converted to
\code{sf}, within which the tessellation will be performed.}

\item{n}{Desired number of Voronoi tiles; as long as \code{poly} is
contiguous (see \code{warn_multipart}), the function will
always return exactly this number of tiles.}

\item{warn_multipart}{Logical. Set \code{FALSE} to suppress the warning about
multipart geometries producing an inconsistent number
of tiles.}
}
\value{
An \code{sf} object containing polygon geometries and an integer unique \code{id}.
}
\description{
\code{voronoi} generates \code{n} Voronoi tiles (also known as Thiessen polygons)
within a polygon using Delaunay triangulation. It is essentially a wrapper
for \code{\link[deldir:deldir]{deldir::deldir()}} which outputs a tidy \code{sf} object.
}
\examples{
frame <- sf::st_sf(rpolygon())
tiles <- voronoi(frame, 50)
plot(tiles)

}
