% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clip_edges}
\alias{clip_edges}
\title{Clip edge polygons}
\usage{
clip_edges(polygons)
}
\arguments{
\item{polygons}{Am `sf`` object.}
}
\value{
An \code{sf} object.
}
\description{
\code{clip_edges()} takes an sf object and removes polygons that touch its bounding
box.
}
