% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{gridded}
\alias{gridded}
\title{Gridded survey units}
\usage{
gridded(frame, n = NULL, size = NULL, orientation = 0)
}
\arguments{
\item{orientation}{}
}
\description{
Generates a regular, rectangular grid of survey units over the frame.
}
\examples{

}
